+++ spear_station/package.xml
@@ -22,6 +22,7 @@
   <!-- Dependencies from nodes -->
   <depend>control_msgs</depend>
   <depend>geometry_msgs</depend>
+  <depend>spear_util</depend>
   <depend>tf2_geometry_msgs</depend>
   <depend>tf2_ros</depend>
   <depend>trajectory_msgs</depend>
